chore: rename pagination for fixing typog
