+++ src/web/src/constants/index.ts
@@ -1,2 +1,2 @@
 export * from './env';
-export * from './defaultPaginiation';
+export * from './defaultPagination';
